+++ src/routes/(tools)/document-generator/Componants/Navbar.svelte
@@ -1,0 +1,21 @@
+<script>
+  
+</script>
+
+<div class="w-1/5 h-auto pt-2 pb-2">
+    <nav class="w-full">
+        <ul class="flex flex-col justify-center items-center gap-4">
+            <li class="w-36 p-2 text-center font-bold rounded-full bg-blue-200 text-black hover:bg-blue-600 hover:text-white cursor-pointer transition duration-300">Assessment</li>
+            <li class="w-36 p-2 text-center font-bold rounded-full bg-blue-200 text-black hover:bg-blue-600 hover:text-white cursor-pointer transition duration-300">Business Prop.</li>
+            <li class="w-36 p-2 text-center font-bold rounded-full bg-blue-200 text-black hover:bg-blue-600 hover:text-white cursor-pointer transition duration-300">Employment Form</li>
+            <li class="w-36 p-2 text-center font-bold rounded-full bg-blue-200 text-black hover:bg-blue-600 hover:text-white cursor-pointer transition duration-300">Letterhead</li>
+            <li class="w-36 p-2 text-center font-bold rounded-full bg-blue-200 text-black hover:bg-blue-600 hover:text-white cursor-pointer transition duration-300">Report</li>
+            <li class="w-36 p-2 text-center font-bold rounded-full bg-blue-200 text-black hover:bg-blue-600 hover:text-white cursor-pointer transition duration-300">Resume</li>
+            <li class="w-36 p-2 text-center font-bold rounded-full bg-blue-200 text-black hover:bg-blue-600 hover:text-white cursor-pointer transition duration-300">Others</li>
+        </ul>
+    </nav>
+</div>
+
+<style>
+  
+</style>
